+++ ChangeLog.md
@@ -5,3 +5,5 @@
 0.1.4.0 - Added curved tracks and track building functionality
 0.1.5.0 - Added hills and interpolation functions
 0.1.5.1 - Added pitched tracks
+0.2.0.1 - Added projecting functionality for road objects
+          Refactored project architecture